requirements.txt
boto==2.34.0
commandr==1.3.2
Fabric==1.10.0
Flask-RESTful==0.3.1
gevent==1.0.1
Jinja2==2.7.3
kazoo==2.0
librato-metrics==0.7.0
raven==5.1.1
requests==2.5.1
